+++ miden-vm/src/repl/mod.rs
@@ -318,8 +318,7 @@ fn execute(
     let advice_inputs = AdviceInputs::default();
     let mut host = DefaultHost::default();
     for library in provided_libraries {
-        host.load_mast_forest(library.mast_forest().clone())
-            .map_err(|err| format!("{err}"))?;
+        host.load_library(library.mast_forest()).map_err(|err| format!("{err}"))?;
     }
 
     let state_iter = miden_processor::execute_iter(