+++ miden-vm/src/tools/mod.rs
@@ -65,8 +65,7 @@ impl Analyze {
         // fetch the stack and program inputs from the arguments
         let stack_inputs = input_data.parse_stack_inputs().map_err(Report::msg)?;
         let advice_inputs = input_data.parse_advice_inputs().map_err(Report::msg)?;
-        let mut host = DefaultHost::default();
-        host.load_mast_forest(StdLibrary::default().mast_forest().clone())?;
+        let host = DefaultHost::default().with_library(&StdLibrary::default())?;
 
         let execution_details: ExecutionDetails =
             analyze(&program, stack_inputs, advice_inputs, host, source_manager)